+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.dto.ReservationCreateRequest;
-import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.dto.ReservationUpdateRequest;
-import roomescape.reservation.dto.ReservationsResponse;
+import roomescape.reservation.controller.dto.ReservationCreateRequest;
+import roomescape.reservation.controller.dto.ReservationResponse;
+import roomescape.reservation.controller.dto.ReservationUpdateRequest;
+import roomescape.reservation.controller.dto.ReservationsResponse;
 import roomescape.reservation.service.ReservationService;
 
 import java.util.List;
