+++ src/main/java/roomescape/reservation/controller/dto/ReservationsResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.controller.dto;
 
 
 import java.util.List;
