+++ src/main/java/roomescape/reservationtime/controller/ReservationTimeAdminController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.reservationtime.domain.ReservationTime;
-import roomescape.reservationtime.dto.ReservationTimeCreateRequest;
-import roomescape.reservationtime.dto.ReservationTimeResponse;
+import roomescape.reservationtime.controller.dto.ReservationTimeCreateRequest;
+import roomescape.reservationtime.controller.dto.ReservationTimeResponse;
 import roomescape.reservationtime.service.ReservationTimeService;
 
 import static org.springframework.http.HttpStatus.CREATED;
