+++ src/main/java/roomescape/reservationtime/controller/dto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime.dto;
+package roomescape.reservationtime.controller.dto;
 
 import roomescape.reservationtime.domain.ReservationTime;
