+++ src/main/java/roomescape/reservationtime/controller/dto/ReservationTimesResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime.dto;
+package roomescape.reservationtime.controller.dto;
 
 import java.util.List;
