+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.theme.dto.ThemeResponse;
-import roomescape.theme.dto.ThemesResponse;
+import roomescape.theme.controller.dto.ThemeResponse;
+import roomescape.theme.controller.dto.ThemesResponse;
 import roomescape.theme.service.ThemeService;
 
 import java.time.LocalDate;
